It turns out after considerations https://stackoverflow.com/questions/77748906/in-a-svelte-writable-store-how-can-i-set-a-different-type-for-getter-setter-in/77750108#77750108, that is useless to over-complicate the type system, when we could easily use a optional `id?`.
I got multiple times in an infinite loop. for now, I rechanged the type system, by using an optional id? because it turns out that even If the user setted his id... in subscribe/set I just replace his id. with this approach If I don't set id typescript doesn't complain, if I set id typescript doesn't complain, and I also have intellisense. for now, I used the length to not go in an infinite loop. basically, when the length changes I do the logic. still a bad approach, but is a work in progress. I have to used this approach since the array is inside a nested objects/array